+++ app/models/transfer/project.rb
@@ -1,0 +1,5 @@
+class Transfer::Project < Project
+  def self.policy_class
+    ProjectPolicy
+  end
+end