+++ db/migrate/20230607103526_create_transfer_tasks_data.rb
@@ -1,0 +1,7 @@
+class CreateTransferTasksData < ActiveRecord::Migration[7.0]
+  def change
+    create_table :transfer_tasks_data, id: :uuid do |t|
+      t.timestamps
+    end
+  end
+end