+++ spec/factories/transfer/tasks_data_factory.rb
@@ -1,0 +1,4 @@
+FactoryBot.define do
+  factory :transfer_tasks_data, class: "Transfer::TasksData" do
+  end
+end