+++ spec/models/transfer/project_spec.rb
@@ -1,0 +1,9 @@
+require "rails_helper"
+
+RSpec.describe Transfer::Project do
+  describe ".policy_class" do
+    it "returns the correct policy" do
+      expect(described_class.policy_class).to eql(ProjectPolicy)
+    end
+  end
+end