+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - The Opening section is now included in all projects, it shows conversion that
   are due to open in the next calendar month and those that were going to open
   but are no longer.
+- Add bare bones Transfer::Project and Transfer::TasksData models
 
 ### Changed